+++ src/app/news/resolvers/news.resolver.ts
