+++ src/app/profile/resolvers/profile-resolvers.module.ts
@@ -1,4 +1,4 @@
 import { NgModule } from '@angular/core';
 
 @NgModule()
-export class ProfileResolvesModule { }
+export class ProfileResolversModule { }
